# Improve get/set performance in HashSet<T>