--- a/src/System.Collections/src/System/Collections/Generic/HashSet.cs
+++ b/src/System.Collections/src/System/Collections/Generic/HashSet.cs
@@ -484,7 +484,8 @@ namespace System.Collections.Generic
             _comparer = (IEqualityComparer<T>)_siInfo.GetValue(ComparerName, typeof(IEqualityComparer<T>))!;
             _freeList = -1;
 
-            if (_comparer == EqualityComparer<T>.Default)
+            // "==" would fail since they're different instances
+            if (Equals(_comparer, EqualityComparer<T>.Default))
             {
                 _comparer = null;
             }