@JeffreyZhao 2019-08-07
The check failed at "Send to Helix". Is it a glitch on check server? Can we re-run the checks?