@danmosemsft 2019-08-07
You should be able to see the test issue by clicking through the log. It's apparently unrelated Http failure. I'll open issue.