@danmosemsft 2019-08-07
https://github.com/dotnet/corefx/issues/40115