@danmosemsft 2019-08-07
It would not be part of this change, but another pending backport to HashSet is https://github.com/dotnet/coreclr/pull/23591.