@JeffreyZhao 2019-08-08
> You should be able to see the test issue by clicking through the log. It's apparently unrelated Http failure. I'll open issue.

Thanks, I've noticed that now. BTW why it's not part of previous step (compile and test) but "Send to Helix", and what's that step?

> Thanks for hte contribution. Do we have sufficient micro benchmark coverage in https://github.com/dotnet/performance ? If not would you consider improving coverage there? This is how we typically measure and prevent regression.

It seems we don't have any coverage for `HashSet<>` and there's only one case for `Dictionary<int, int>.ContainsValue`, which is a bit strange to me since it's one of the least used methods. Yes, I can add some for `HashSet<>` and probably `Dictionary<,>` too.

> It would not be part of this change, but another pending backport to HashSet is dotnet/coreclr#23591.

Yes, I've noticed that too. That's what I mean by this:

> I've intentionally kept the origin implementation and comments as
> much as possible, and the optimized code is not exactly the same as
> in Dictionary<T>. Further micro-optimizations might be possible
> but I think it's cleaner to keep it that way in the first step.

There're actually more optimizations can be put in `HashSet<>` like the one you mentioned, and also including the changes to other `Set` methods. My thought is to do it gradually so it's easier to review, test and merge. Those optimizations could be made later. Now I think it could be made after having more performance coverage.