@safern 2019-08-08
>the constructors and the OnDeserialization
method have been modified accordingly to save null instead of the
EqualityComparer<T>.Default if it's been passed/saved.

@ViktorHofer would this break serialization scenarios?