@ViktorHofer 2019-08-08
> @ViktorHofer would this break serialization scenarios?

No it doesn't break serialization as during serialization the default equality comparer is used if the comparer is null. The description is a bit misleading but the change is fine.