@JeffreyZhao 2019-10-15
> Overall, looks good. Are you including benchmark tests into https://github.com/dotnet/performance/tree/master/src/benchmarks?
> I'd like to get @stephentoub input on this as well.

@safern I'll move my benchmark to performance project when I finish learning how to do it properly.