@JeffreyZhao 2019-10-16
> Overall, looks good. Are you including benchmark tests into https://github.com/dotnet/performance/tree/master/src/benchmarks?

@safern I just checked the existing benchmarks, and it seems already covers the `Contains` and `Add` cases. Do I still need to add my benchmarks to that project?

```
> dotnet run -c Release -f netcoreapp3.0 --filter System.Collections.*.HashSet --list tree
System
 └─Collections
    ├─AddGivenSize<Int32>
    │  └─HashSet
    ├─AddGivenSize<String>
    │  └─HashSet
    ├─ContainsFalse<Int32>
    │  └─HashSet
    ├─ContainsFalse<String>
    │  └─HashSet
    ├─ContainsTrue<Int32>
    │  └─HashSet
    ├─ContainsTrue<String>
    │  └─HashSet
    ...
```