@safern 2019-10-16
> @safern I just checked the existing benchmarks, and it seems already covers the Contains and Add cases. Do I still need to add my benchmarks to that project?

You're right. That should be sufficient then.

Do you have the latest benchmark results with the current state? Thanks in advance.