@JeffreyZhao 2019-10-18
@danmosemsft @safern Find some more information regarding `[IterationSetup]` here: https://github.com/dotnet/BenchmarkDotNet/issues/1003#issuecomment-452767170

> When you use [IterationSetup] BDN executes the benchmark once per iteration. If this a very short operation, the result is unstable. This is why we produce this warning.

Anyway, I've added some more benchmarks in my fork of dotnet/performance: https://github.com/JeffreyZhao/performance/tree/Collection-Remove-Benchmarks

I've added two new set of benchmarks (`RemoveFalse` and `CopyAndRemove`), mainly for `HashSet<>` and `Dictionary<>`, and I've pasted the result below.

I've also found the method call of `InternalGetHashCode` hasn't been inlined in `Remove` ([thanks SharpLab!](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgEYBYAKGIAYACY8lAbhvqfIDoAlAVwB2GAJb4Y3AMIR8ABxEAbGFADKygG4iwMXO1qMVAC2xRZAGWzA+Q0eL00A2gCkRGAOIxByrQAoMAT1kYCAAzHxFhAEpIgF0OAGYmUgYAFR0MAAk8QzUMAB4UgD4aAG8aBgqGWSgRdWwMGAZIQVwMBgi2swgAd2V48gAhVwBZPABrBgBeBjoEAHYAMSXlhb0Gcsrq2vrGjocYgH4GAH1gfjAxmAxdDYqtuoaGFQUIDH2T3BfrqYYAExgQth+AoMABCVgMAD0kNSAHkACKwuAAOQAqmYzABBAZmACiIFShhEuAYuEMEGBvwYwEaz1e+wOaGp/DaGEMjVguGBokEAHMGNhBIJXvURBAWk0gbgdAwPIIAAomUTYBQLbBgbD/FQUqDaBgYCDUl7dBj8WQE5RQaAMAmowSWJT6w0a1oMCDAABWMDAbVC+sCjQA5MMtFbcKEMFIIP9MfaFP5cMSpEYTLJuCp/MJsAgo3JFKLxXbXBmswhA07/UEGMHQxBwyFI9IY3GE0nJCnTOnMxhs9wBngYCkA/DvQoTAXBCWe2XblUag9dsITpAbGtNvOdu0l8cx60AJKCf4INd3DePDonEKwGBmYkYE9z7aPPe4gCO/BVrn80jkJmUBUKI5jkgX9YCgB97k3C9jnUZRE3FexqEqJhEmACAIAUBheBgfAIFgnwUnaBp8EiWcyiQ5DKhEEIGB8U5zkub5QWmQRgQUUiKMohhyK4riX3fT8Ah/WQ/ygACjhAkSwJ+YCZCk5QH14rc2mMMkm0aaYDwaKA4w8TJsnU8JiKZSTRMiRTeIvM4LiuH5VMMdSGAAUhOazGNwbgzE8Xk2QsriL13Nppjgcg/Moi9IAUBRiTFQRpBsH46DC5C6TeGJSS+ElpmOT5XhuTilIYEJrXCJcRBktyrlwBxKowdK4AYUL2gYQppiShhAp+EQmXK6ZcuuBwRBibgvAQDAOMKyoeMm8KaJ8frqqG7h7McyZphW6NGgAMi2hgtOUOM3w/BRcHmzLBuGuoFH4GAeuMpo5LMiaZuQ6aXv8ubOryGZnveyi3r+3joSKkQoFdVxlAnCEzV+TdatwWdAco+ibIG2r0r686ltGtoAGpGuSmaAF9EcBmATpgUm/oBpHkOBrlgGld9PDaCGoAnXRTVkWHHkDHHy1kCAOjgqnaYWhxAuGnGfnF7GYDGwnJpJgqkdl4aNv+H4QsVwrqNogRhDEGAMnJoIwe4PdcGwkJlE8bRYSgaQswiK2AVtwRtFwACfGiUWXpp2mKjV7grpun5/kBblwT94mY8m4Ppeyq8YBvO9EMDipZJsOA4B1pSYLg2LcdxvPLLmrOlzWn648KgPA53PAMAPI9EtLybjmT1PXWC0Ka6U5WM4YcnpT73i69pjvr1vbv2jbyiB4z4g5n1KAbrn5XR4qPWfEi6L4Liikl1ajK8s87y2V+97x5e4GUnZJpjAiN0aJZmoZWKqB8BJbAOvQ2QmW6ESMAhgGDiEFN/B6Ht+BQFgEuGGm5VIMGMLIIIXhfjcE3pRNkVoTReBNAeK6IhfiwjNhOXECBtCyFEOKHwAAiFQvALaCEIcQ0h1DBDHCdmAaBsCMAkMhuw3AyJXgqDNILKADRfi0PMpghef1d4xXFPFYQxcdZyPnqLYGMBXDsigK5BiVV2gklYlFN0ejuh4EEIGNoxUhC/FFkvIqKppQPnUbOYWOkVSkgwKvH0TwvhkVFh4uMykkEGU2hCSiwMui9D0f0akrgSR+nsg9f4TIQrtBokIfg0p7EqyosIA6XiLw40iUpYGzd5bPwYNLV+/h0nkEyR1RuQTCmeMwoRUOMBXGzlnMDBwwwrjkl+HuOQCgfCDLZNGUZsgFAkMEdwTEvJeSckTLBA80VBARF5LEWckFHjMCQKE/a7S9JZDUptAiREcJMn4sdL8wlRLiQeqBZQl9x7bwhvgKYLE2KX1ejXRxSVRYb3yVvJcGsNIvPklAA43AznhP+EZHCkQGAHCOF8+FVxzkOUuTIsFTBl6QoYLtGJfRBgjHGD06gfSYQDKGdMsZEyGUjLGfM2KHklkrJ0GsmAGyIjbN2QVfZjQ0IYT2m0w6AkTpXIQEyQi9S9pHUEt+R6YFnmmTAu8hxy9NXKDhcqmVcqGD+FReihghqHlqv/EUbgI5I4gm4Ia06xrTXUqJkAA=)) so I've submitted another commit to put `AggressiveInlining` on `InternalGetHashCode` and `InternalEquals` methods. In `Dictionary<T>` it's manually inlined but I've extracted a helper method for less code duplications.

I've made three builds of CoreFx for benchmark:

1. netcoreapp-original: Original implementation without this pull request.
2. netcoreapp-hashset: `HashSet` optimization without `InternalGetHashCode` inlining.
3. netcoreapp-hashset-inline: `HashSet` optimization with `InternalGetHashCode` inlining.

```
dotnet run -c Release -f netcoreapp3.0 --filter "*Remove*.HashSet" --join --coreRun \
    C:\Projects\dotnet\netcoreapp-original\coreRun.exe
    C:\Projects\dotnet\netcoreapp-hashset\coreRun.exe
    C:\Projects\dotnet\netcoreapp-hashset-inline\coreRun.exe
```
<details>
<summary>Expand to see benchmark results.</summary>

```ini
BenchmarkDotNet=v0.11.5.1159-nightly, OS=Windows 10.0.17763.805 (1809/October2018Update/Redstone5)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100
  [Host]     : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), X64 RyuJIT
  Job-QFCKJK : .NET Core ? (CoreCLR 5.0.19.38002, CoreFX 5.0.19.51701), X64 RyuJIT
  Job-TMFJJO : .NET Core ? (CoreCLR 5.0.19.38002, CoreFX 5.0.19.51001), X64 RyuJIT
  Job-FUBAAA : .NET Core ? (CoreCLR 5.0.19.38002, CoreFX 5.0.19.51701), X64 RyuJIT
```

|                       Type |  Method |                              Toolchain | Size |      Mean |     Error |    StdDev |    Median |       Min |       Max | Ratio | RatioSD |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |-------- |--------------------------------------- |----- |----------:|----------:|----------:|----------:|----------:|----------:|------:|--------:|--------:|------:|------:|----------:|
|       CopyAndRemove&lt;Int32&gt; | HashSet | \netcoreapp-hashset-inline\coreRun.exe |  512 |  8.050 us | 0.1281 us | 0.1198 us |  8.004 us |  7.881 us |  8.288 us |  0.78 |    0.01 |  4.5778 |     - |     - |    8456 B |
|       CopyAndRemove&lt;Int32&gt; | HashSet |        \netcoreapp-hashset\coreRun.exe |  512 |  9.102 us | 0.0761 us | 0.0674 us |  9.102 us |  8.999 us |  9.198 us |  0.88 |    0.01 |  4.5872 |     - |     - |    8456 B |
|       CopyAndRemove&lt;Int32&gt; | HashSet |       \netcoreapp-original\coreRun.exe |  512 | 10.296 us | 0.1012 us | 0.0897 us | 10.303 us | 10.154 us | 10.449 us |  1.00 |    0.00 |  4.5569 |     - |     - |    8456 B |
|                            |         |                                        |      |           |           |           |           |           |           |       |         |         |       |       |           |
|      CopyAndRemove&lt;String&gt; | HashSet | \netcoreapp-hashset-inline\coreRun.exe |  512 | 32.356 us | 0.3855 us | 0.3606 us | 32.289 us | 31.874 us | 33.163 us |  1.12 |    0.01 |  3.4022 |     - |     - |   10536 B |
|      CopyAndRemove&lt;String&gt; | HashSet |        \netcoreapp-hashset\coreRun.exe |  512 | 31.714 us | 0.2834 us | 0.2512 us | 31.795 us | 31.297 us | 32.074 us |  1.10 |    0.01 |  3.3886 |     - |     - |   10536 B |
|      CopyAndRemove&lt;String&gt; | HashSet |       \netcoreapp-original\coreRun.exe |  512 | 28.788 us | 0.3398 us | 0.3178 us | 28.656 us | 28.427 us | 29.370 us |  1.00 |    0.00 |  3.4153 |     - |     - |   10536 B |
|                            |         |                                        |      |           |           |           |           |           |           |       |         |         |       |       |           |
|         RemoveFalse&lt;Int32&gt; | HashSet | \netcoreapp-hashset-inline\coreRun.exe |  512 |  3.936 us | 0.0349 us | 0.0326 us |  3.928 us |  3.890 us |  4.005 us |  0.61 |    0.01 |       - |     - |     - |         - |
|         RemoveFalse&lt;Int32&gt; | HashSet |        \netcoreapp-hashset\coreRun.exe |  512 |  5.583 us | 0.1057 us | 0.0989 us |  5.559 us |  5.434 us |  5.748 us |  0.87 |    0.02 |       - |     - |     - |         - |
|         RemoveFalse&lt;Int32&gt; | HashSet |       \netcoreapp-original\coreRun.exe |  512 |  6.409 us | 0.0622 us | 0.0520 us |  6.402 us |  6.314 us |  6.508 us |  1.00 |    0.00 |       - |     - |     - |         - |
|                            |         |                                        |      |           |           |           |           |           |           |       |         |         |       |       |           |
|        RemoveFalse&lt;String&gt; | HashSet | \netcoreapp-hashset-inline\coreRun.exe |  512 | 18.432 us | 0.0971 us | 0.0861 us | 18.421 us | 18.277 us | 18.547 us |  0.88 |    0.01 |       - |     - |     - |         - |
|        RemoveFalse&lt;String&gt; | HashSet |        \netcoreapp-hashset\coreRun.exe |  512 | 18.432 us | 0.1500 us | 0.1403 us | 18.413 us | 18.168 us | 18.702 us |  0.88 |    0.01 |       - |     - |     - |         - |
|        RemoveFalse&lt;String&gt; | HashSet |       \netcoreapp-original\coreRun.exe |  512 | 20.980 us | 0.1797 us | 0.1501 us | 20.969 us | 20.669 us | 21.303 us |  1.00 |    0.00 |       - |     - |     - |         - |
|                            |         |                                        |      |           |           |           |           |           |           |       |         |         |       |       |           |
|  CreateAddAndRemove&lt;Int32&gt; | HashSet | \netcoreapp-hashset-inline\coreRun.exe |  512 | 18.468 us | 0.3007 us | 0.2813 us | 18.482 us | 18.024 us | 19.119 us |  0.76 |    0.02 |  8.8038 |     - |     - |   27712 B |
|  CreateAddAndRemove&lt;Int32&gt; | HashSet |        \netcoreapp-hashset\coreRun.exe |  512 | 20.677 us | 0.2535 us | 0.2371 us | 20.721 us | 20.226 us | 20.996 us |  0.85 |    0.01 |  8.8120 |     - |     - |   27712 B |
|  CreateAddAndRemove&lt;Int32&gt; | HashSet |       \netcoreapp-original\coreRun.exe |  512 | 24.407 us | 0.3310 us | 0.2934 us | 24.458 us | 23.820 us | 24.916 us |  1.00 |    0.00 |  8.7461 |     - |     - |   27712 B |
|                            |         |                                        |      |           |           |           |           |           |           |       |         |         |       |       |           |
| CreateAddAndRemove&lt;String&gt; | HashSet | \netcoreapp-hashset-inline\coreRun.exe |  512 | 60.367 us | 1.1194 us | 1.0471 us | 59.980 us | 59.257 us | 62.924 us |  0.96 |    0.02 | 10.8902 |     - |     - |   34481 B |
| CreateAddAndRemove&lt;String&gt; | HashSet |        \netcoreapp-hashset\coreRun.exe |  512 | 59.649 us | 0.4124 us | 0.3444 us | 59.722 us | 59.007 us | 60.182 us |  0.95 |    0.01 | 10.8173 |     - |     - |   34480 B |
| CreateAddAndRemove&lt;String&gt; | HashSet |       \netcoreapp-original\coreRun.exe |  512 | 62.911 us | 0.7728 us | 0.7228 us | 62.884 us | 61.911 us | 64.280 us |  1.00 |    0.00 | 10.7932 |     - |     - |   34480 B |

</details>