@JeffreyZhao 2019-10-25
(Previous comment removed as I realized I'm running with a different data set.)

Today's runs of `CopyAndRemove<String>.HashSet` has different results.

Here's three consecutive runs:

```
dotnet run -c Release -f netcoreapp3.0 --filter "*CopyAndRemove<String>.HashSet" --join --coreRun \
    C:\Projects\dotnet\netcoreapp-original\coreRun.exe \
    C:\Projects\dotnet\netcoreapp-hashset\coreRun.exe
```

<details>
<summary>Expand to see benchmark results.</summary>

|  Method |                        Toolchain | Size |     Mean |    Error |   StdDev |   Median |      Min |      Max | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |--------------------------------- |----- |---------:|---------:|---------:|---------:|---------:|---------:|------:|--------:|-------:|------:|------:|----------:|
| HashSet |  \netcoreapp-hashset\coreRun.exe |  512 | 26.75 us | 0.526 us | 0.492 us | 26.65 us | 25.95 us | 27.71 us |  0.93 |    0.02 | 3.4781 |     - |     - |  10.29 KB |
| HashSet | \netcoreapp-original\coreRun.exe |  512 | 28.63 us | 0.357 us | 0.334 us | 28.65 us | 28.11 us | 29.37 us |  1.00 |    0.00 | 3.4722 |     - |     - |  10.29 KB |
                                                                                                                                                                              
|  Method |                        Toolchain | Size |     Mean |    Error |   StdDev |   Median |      Min |      Max | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |   
|-------- |--------------------------------- |----- |---------:|---------:|---------:|---------:|---------:|---------:|------:|--------:|-------:|------:|------:|----------:|   
| HashSet |  \netcoreapp-hashset\coreRun.exe |  512 | 27.02 us | 0.398 us | 0.372 us | 26.99 us | 26.58 us | 27.74 us |  0.96 |    0.02 | 3.4722 |     - |     - |  10.29 KB |   
| HashSet | \netcoreapp-original\coreRun.exe |  512 | 28.22 us | 0.310 us | 0.275 us | 28.16 us | 27.94 us | 28.95 us |  1.00 |    0.00 | 3.4537 |     - |     - |  10.29 KB |   
                                                                                                                                                                                 
|  Method |                        Toolchain | Size |     Mean |    Error |   StdDev |   Median |      Min |      Max | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |--------------------------------- |----- |---------:|---------:|---------:|---------:|---------:|---------:|------:|--------:|-------:|------:|------:|----------:|
| HashSet |  \netcoreapp-hashset\coreRun.exe |  512 | 26.76 us | 0.400 us | 0.374 us | 26.67 us | 26.12 us | 27.48 us |  0.94 |    0.02 | 3.4606 |     - |     - |  10.29 KB |
| HashSet | \netcoreapp-original\coreRun.exe |  512 | 28.55 us | 0.223 us | 0.209 us | 28.58 us | 28.21 us | 29.00 us |  1.00 |    0.00 | 3.4594 |     - |     - |  10.29 KB |

</details>

We can see that the optimized version is always faster.